INT-2134 addressed the latest PR comments made 'expireGroupsUponCompletion' attribute protected and moved its setter to the AggregatingMessageHandler, added '!this.shouldExpireGroupsUponCompletion()' check to the ResequencingMessageHandler.cleanUpForReleasedGroup(..) method, added IF (expireGroupsUponCompletion) to setExpireGroupsUponCompletion(..)
